@@ -2,13 +2,14 @@
   "name": "fastify-secrets-core",
   "version": "3.0.0",
   "description": "Simplify development of fastify-secrets plugins",
+  "type": "module",
   "main": "lib/fastify-secrets-core.js",
   "scripts": {
     "lint": "eslint lib/**/*.js test/**/*.js",
     "lint:fix": "npm run lint -- --fix",
     "lint:staged": "lint-staged",
-    "test": "tap --no-browser --coverage-report=html --coverage-report=text  test",
-    "test:ci": "tap --no-color --coverage-report=json --coverage-report=text test"
+    "test": "c8 node --test test/*.test.js --check-coverage --statements 97.67 --branches 85.71 --functions 100 --lines 97.67",
+    "test:ci": "c8 node --test test/*.test.js --check-coverage --statements 97.67 --branches 85.71 --functions 100 --lines 97.67"
   },
   "repository": {
     "type": "git",
@@ -28,27 +29,25 @@
   },
   "homepage": "https://github.com/nearform/fastify-secrets-core#readme",
   "engines": {
-    "node": ">= 12.13.0"
+    "node": ">=22"
   },
   "dependencies": {
     "fastify-plugin": "^5.0.1",
-    "p-map": "^4.0.0",
-    "p-props": "^4.0.0"
+    "p-map": "^7.0.3",
+    "p-props": "^6.0.0"
   },
   "devDependencies": {
+    "c8": "^10.1.3",
     "eslint": "^8.18.0",
-    "eslint-config-prettier": "^10.0.1",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.22.1",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-n": "^16.0.2",
-    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-prettier": "^5.5.1",
     "eslint-plugin-promise": "^6.0.0",
-    "husky": "^9.0.11",
-    "lint-staged": "^16.0.0",
-    "prettier": "^3.0.1",
-    "proxyquire": "^2.1.3",
-    "sinon": "^21.0.0",
-    "tap": "^16.0.0"
+    "husky": "^9.1.7",
+    "lint-staged": "^16.1.2",
+    "prettier": "^3.6.2"
   },
   "lint-staged": {
     "*.js": [
