@@ -1,13 +1,11 @@
-'use strict'
-
-const fp = require('fastify-plugin')
-const pProps = require('p-props')
-const pMap = require('p-map')
+import fp from 'fastify-plugin'
+import pMap from 'p-map'
+import pProps from 'p-props'
 
 const DEFAULT_GET_CONCURRENCY = 5
 
 function assertOptions(opts) {
-  if (!opts || !opts.secrets) {
+  if (!opts?.secrets) {
     throw new Error('fastify-secrets: no secrets requested')
   }
 }
@@ -19,7 +17,7 @@ function assertPluginAlreadyRegistered(fastify, opts) {
     throw new Error('fastify-secrets has already been registered')
   }
 
-  if (namespace && fastify.secrets && fastify.secrets[namespace]) {
+  if (namespace && fastify.secrets?.[namespace]) {
     throw new Error(`fastify-secrets '${namespace}' instance name has already been registered`)
   }
 }
@@ -65,7 +63,7 @@ async function refresh(client, fastify, opts, refs) {
   return refreshedSecrets
 }
 
-function buildPlugin(Client, pluginOpts) {
+function buildPlugin(Client, pluginOpts, buildOpts = { fp }) {
   async function FastifySecretsPlugin(fastify, opts) {
     assertOptions(opts)
     assertPluginAlreadyRegistered(fastify, opts)
@@ -89,7 +87,7 @@ function buildPlugin(Client, pluginOpts) {
     }
   }
 
-  const plugin = fp(FastifySecretsPlugin, {
+  const plugin = buildOpts.fp(FastifySecretsPlugin, {
     fastify: '5.x',
     ...pluginOpts
   })
@@ -98,4 +96,5 @@ function buildPlugin(Client, pluginOpts) {
   return plugin
 }
 
-module.exports = buildPlugin
+export default buildPlugin
+export { buildPlugin }
