@@ -1,7 +1,4 @@
-'use strict'
+import buildPlugin from './build-plugin.js'
 
-const buildPlugin = require('./build-plugin')
-
-module.exports = {
-  buildPlugin
-}
+export default buildPlugin
+export { buildPlugin }
